requirements.txt
streamlit
groq
python-dotenv
sentence-transformers
numpy
pypdf
scikit-learn